requirements.txt
# Core dependencies
apache-airflow==2.7.3
kafka-python==2.0.2
pyspark==3.4.1
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Monitoring
prometheus-client==0.17.1
grafana-api==1.0.3

# Utilities
requests==2.31.0
pyyaml==6.0.1

